+++ src/onegov.feriennet/onegov/feriennet/forms/attendee.py
@@ -319,11 +319,11 @@ class AttendeeSignupForm(AttendeeBase):
             attendees = AttendeeCollection(self.request.session)
             birth_date = attendees.by_id(self.attendee.data).birth_date
 
-        if self.model.is_too_old(birth_date):
+        if self.model.is_too_old(birth_date, wiggle_room=0):
             self.attendee.errors.append(_("The attendee is too old"))
             return False
 
-        if self.model.is_too_young(birth_date):
+        if self.model.is_too_young(birth_date, wiggle_room=0):
             self.attendee.errors.append(_("The attendee is too young"))
             return False